Fix issue where priming large messages would cause a 500
The request would fail when priming a large json payload.  This
appears to be due to the post being split across multiple packets
and the content-type not being set.  So I'm guessing sinatra
could not put the data back together again without some hint of
what it was.

For now just doing a basic check for json and then defaulting to
xml (like the normal response mechanism).